+++ src/main/java/data_bases/hw2/task_03/entities/Transportable.java
@@ -2,10 +2,10 @@ package data_bases.hw2.task_03.entities;
 
 public interface Transportable {
     Double costDelivery();
-    Double getWidth();
-    Double getHeight();
-    Double getLength();
-    default Double getVolume(){
+    Integer getWidth();
+    Integer getHeight();
+    Integer getLength();
+    default Integer getVolume(){
         return getWidth() * getHeight() * getLength();
     }
 }
